move SemesterProgramSelector into _compoonents dir
