+++ frontend/src/routes/events/semesterprogramm/_components/SemesterProgramSelector.svelte
