+++ frontend/src/routes/events/semesterprogramm/sose24/+page.svelte
@@ -2,7 +2,7 @@
 	import header from '$lib/assets/pages/events/semesterprogramm/herzenssache_21_9.jpg';
 	import HeroShot from '$lib/components/HeroShot.svelte';
 	import type { PageData } from '../$types';
-	import SemesterProgramSelector from '../SemesterProgramSelector.svelte';
+	import SemesterProgramSelector from '../_components/SemesterProgramSelector.svelte';
 	import dayjs from 'dayjs';
 
 	export let data: PageData;
