+++ frontend/src/routes/events/semesterprogramm/sose25/+page.svelte
@@ -3,7 +3,7 @@
 	import HeroShot from '$lib/components/HeroShot.svelte';
 	import dayjs from 'dayjs';
 	import type { PageData } from './$types';
-	import SemesterProgramSelector from '../SemesterProgramSelector.svelte';
+	import SemesterProgramSelector from '../_components/SemesterProgramSelector.svelte';
 
 	export let data: PageData;
 </script>
